Read the post code from the LPC-PCC device
This commit defines a build time option which tells that
post code needs to be read from the snoop port or the PCC.

This commit also changes the logic to process the post code.
On the AMD platfrom, Post code is of 8 bytes and we need to
form 2 byte post code from 8 bytes.

Signed-off-by: Ratan Gupta <ratankgupta31@gmail.com>